Fix wrong type in redmean

+++ shaders/lib/redmean.glsl
@@ -14,7 +14,7 @@ float16_t sq_redmean(f16vec3 a, f16vec3 b) {
 	immut float16_t r = mix(a.r, b.r, float16_t(0.5)) * float16_t(0.99609375);
 	immut f16vec3 d = a - b;
 
-	return dot(d*d, vec3(
+	return dot(d*d, f16vec3(
 		float16_t(2.0) + r,
 		float16_t(4.0),
 		float16_t(2.99609375) - r
